Added Ability to Pre-Check Multiple Administrative Controls
This commit adds the ability to pre-check multiple controls that require administrative privileges. This allows the script to be run as a standard user to get the correct policies, but elevate for 'computer based' controls to get a comple audit.

This change will allow most of the controls that are not being checked currently to be added in the coming weeks.